requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.4
colorama==0.4.4
Django==3.2
future==0.18.2
html5lib==1.1
Pillow==8.2.0
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
qrcode==6.1
reportlab==3.5.67
six==1.15.0
sqlparse==0.4.1
webencodings==0.5.1
xhtml2pdf==0.2.5
